+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/CameraTestUtils.swift
@@ -100,7 +100,8 @@ enum CameraTestUtils {
   /// Creates a test sample buffer.
   /// @return a test sample buffer.
   static func createTestSampleBuffer(
-    timestamp: CMTime = .zero, duration: CMTime = CMTimeMake(value: 1, timescale: 44100)
+    timestamp: CMTime = .zero,
+    duration: CMTime = CMTimeMake(value: 1, timescale: 44100)
   ) -> CMSampleBuffer {
     var pixelBuffer: CVPixelBuffer?
     CVPixelBufferCreate(kCFAllocatorDefault, 100, 100, kCVPixelFormatType_32BGRA, nil, &pixelBuffer)
@@ -130,7 +131,8 @@ enum CameraTestUtils {
   /// Creates a test audio sample buffer.
   /// @return a test audio sample buffer.
   static func createTestAudioSampleBuffer(
-    timestamp: CMTime = .zero, duration: CMTime = CMTimeMake(value: 1, timescale: 44100)
+    timestamp: CMTime = .zero,
+    duration: CMTime = CMTimeMake(value: 1, timescale: 44100)
   ) -> CMSampleBuffer {
     var blockBuffer: CMBlockBuffer?
     CMBlockBufferCreateWithMemoryBlock(